Relax greta timestyle matching regular expression
Relax greta match_expr to match floats with fewer than 6 decimal
places, as the str -> float -> str operations change
'2013055.000000000' to '2013055.0' and then the regular expression
won't match even if greta is specified as the format.  Also, in
greta_to_date, create a string of the float that has sufficient
decimal places to match the original greta regular expression.

Closes #7